Story 1: Changed that all concrete classes implement SStatus, which is an interface. Methods now accordingly return said SStatus. Also refined flow of control in SC to reflect possibility of duplicate details
