--- src/main/java/com/acme/statusmgr/beans/ServerStatusInterface.java
@@ -1,11 +1,0 @@
-package com.acme.statusmgr.beans;
-
-public interface ServerStatusInterface {
-    String getContentHeader();
-    Integer getRequestCost();
-    String getStatusDesc();
-    long getId();
-
-
-
-}
